use global mocks for block storage tests
You can now also pass a file name to the mock and it will use that to
create the returned buffer rather than a hard coded string.

So `getBlob('filename')` will return a buffer of `filename.pdf`

You can also assert on whether the static method was called or not.